pyproject.toml
[tool.poetry]
name = "metadatax"
version = "0.5.0a6"
description = "PAM acquisition metadata"
authors = ["Your Name <you@example.com>"]
license = "GPL-3.0"
readme = "README.md"

[tool.poetry.dependencies]
python = ">=3.9,<4.0"
Django = "^3.2"
django-filter = ">=2"
djangorestframework = "^3.15.1"
python-dotenv = "^1.0.1"
psycopg2-binary = "^2.8.6"
django-model-prefix = "^1.1.0"
graphene-django = "^3.2.3"
graphene-django-pagination = { url = "https://github.com/instruct-br/graphene-django-pagination/archive/master.zip" }
django-admin-multiple-choice-list-filter = "^0.1.1"
django-better-admin-arrayfield = "^1.4.2"
django-durationwidget = "^1.0.5"
graphene-django-optimizer = "^0.10.0"
#django-extension = { path = "../django-extension", develop = true }
django-extension = { url = "https://github.com/Project-OSmOSE/django-extension/releases/download/v0.1.3/django_extension-0.1.3.tar.gz" }

[tool.poetry.group.dev.dependencies]
drf-yasg = "^1.21.7"
drf-spectacular = "^0.27.2"
black = "^22.10.0"

[tool.poetry.group.prod.dependencies]
gunicorn = "^22.0.0"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
